Dimensions First Commit
First commit of a wip dimensions/virtual worlds system for FiveM.
Main use case would be for interiors as it disables firing and sets the alpha of all players to 0 and disables collisions with them if they aren't in the same dimension. Nothing else is changed, such as objects, cars, peds, etc.
It isn't 100% ready, obviously, and sometimes the collisions aren't always disabled or not all the players have the alphas change properly, allowing some to see others in other dimensions.